+++ osu.Game/Screens/Tatakae/TatakaeScreenTab.cs
@@ -1,0 +1,7 @@
+namespace osu.Game.Screens.Tatakae
+{
+    public enum TatakaeScreenTab
+    {
+        General
+    }
+}